refactor: extract version dependent logic to `_compat` module
